doorstop/gui: move the Application class to a separate file
I cannot open the second refs PR because I am hitting the 1000 limit on the
gui/main.py file.

This is the easiest refactoring of the main.py I could think of to relax its
size.

This is only a move type change, no changes to the behavior.
